requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
twilio==8.10.3
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
openai==1.6.1